manager/requirements-core.txt
# Core dependencies for MarchProxy Manager (without SAML)
py4web>=1.20230507.1
pydal>=20230521.1
psycopg2-binary>=2.9.9
redis>=5.0.1
bcrypt>=4.1.2
PyJWT>=2.8.0
pyotp>=2.9.0
qrcode>=7.4.2
python-dotenv>=1.0.0
pydantic>=2.5.3
httpx>=0.25.2
uvicorn>=0.25.0
gunicorn>=21.2.0

# API and web
fastapi>=0.108.0
jinja2>=3.1.2
aiofiles>=23.2.1
python-multipart>=0.0.6

# Certificate management
cryptography>=41.0.7
certifi>=2023.11.17

# Monitoring and logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7
sentry-sdk>=1.39.1

# Utilities
click>=8.1.7
PyYAML>=6.0.1
pendulum>=3.0.0
validators>=0.22.0